Add files via upload
Make both sentences 30% larger. The french sentence is in colibri and letters falls from the sky bounce at the bottom and then install themselves one after the other with fireworks effect. Then the spanish version in code ms font is typed on the screen with bip sounds of different musical keys. when the user tap on the sentences they do special effects randomly and make place for a chatbot greeting the user by Hola! Que Guapo! Como esta? Branch pandora api and fliet with the user. chatbot is fluent in englosh and spanish. He's flirty and learns from experiences